refactor: Turn global_hotkeys into a directory
